`cloneDeep`: use `structuredClone` in platform where that's possible
