+++ .changeset/nice-icons-approve.md
@@ -1,0 +1,5 @@
+---
+"@apollo/client": patch
+---
+
+`cloneDeep`: use `structuredClone` in platform where that's possible
