+++ package.json
@@ -62,7 +62,10 @@
     "./testing/core": "./src/testing/core/index.ts",
     "./testing/experimental": "./src/testing/experimental/index.ts",
     "./testing/react": "./src/testing/react/index.ts",
-    "./utilities": "./src/utilities/index.ts",
+    "./utilities": {
+      "react-native": "./src/utilities/react-native/index.ts",
+      "default": "./src/utilities/index.ts"
+    },
     "./utilities/internal": {
       "production": "./src/utilities/internal/index.production.ts",
       "default": "./src/utilities/internal/index.ts"
