+++ src/utilities/index.react-native.ts
@@ -1,0 +1,2 @@
+export * from "./index.js";
+export { cloneDeep } from "./common/cloneDeep.react-native.js";
